rename package without underscopes
