+++ common/persistence/errorinjectors/injectors_test.go
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package error_injectors
+package errorinjectors
 
 import (
 	"context"
